SMS-4: added calculateEstimatedGroupSize

+++ sms_smpp/src/main/java/org/sakaiproject/sms/api/SmsBilling.java
@@ -126,7 +126,7 @@ public interface SmsBilling {
 	 * 
 	 * @return the double
 	 */
-	public double convertCreditsToAmount(int creditCount);
+	public float convertCreditsToAmount(int creditCount);
 
 	/**
 	 * Insert a new account and return the new account id.
